messaging/requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
async-timeout==3.0.1
brotlipy==0.7.0
cchardet==2.1.6
pyjwt==1.7.1
